// resources/js/types/crud.ts
export interface TableColumn {
    key: string;
    label: ReactNode;
    sortable?: boolean;
    type?: 'text' | 'badge' | 'image' | 'date' | 'currency' | 'boolean' | 'link' | 'custom';
    className?: string;
    linkClassName?: string;
    href?: string | ((row: any) => string);
    openInNewTab?: boolean;
    render?: (value: any, row: any) => React.ReactNode;
  }